% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_NHANES_data.R
\name{get_NHANES_data}
\alias{get_NHANES_data}
\title{get_NHANES_data}
\usage{
get_NHANES_data(codes_file = NULL)
}
\arguments{
\item{codes_file}{Manually created xls or xlsx file containing 3 sheets:  1. NHANES chemicals to include (with identifier,
code, file, demographic, and units), 2. Associated weights, filenames, and column names associated with each phase used,
and 3. Parent-metabolite map containing chemical identifiers and molecular weights.}
}
\value{

}
\description{
Downloads NHANES data files needed for exposure inference based on a manually curated file describing which data to use
}
\examples{


}
