% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_estimates.R
\name{plot_subpops}
\alias{plot_subpops}
\title{plot_subpops}
\usage{
plot_subpops(path_to_lPs = NULL, SUBPOP = "all")
}
\arguments{
\item{path_to_lPs}{String representing the location of the lPsamps outputs.
All subpopulations that are desired in the plot need to
be within the directory provided.}

\item{SUBPOP}{A character vector of the subpopulations to be plotted.  These
must exactly match the subpopulations contained in the original
codes_file.
#@param save_location String giving the path of where to save the output tableand plot.

#@param save_output Logical indicating if plot should be saved.}
}
\value{
exposureDF:  data.frame with rows representing the chemicals and columns
representing the different subpopulations
p:  The resulting plot
}
\description{
Plot the median exposure estimate for all subpopulations of each parent
chemical that have calculated exposure estimates.  Sorted from highest
to lowest predicted exposure.
}
