% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{rcreatinine}
\alias{rcreatinine}
\title{rcreatinine}
\usage{
rcreatinine(newdata)
}
\arguments{
\item{newdata}{A data.frame with individual-specific data including the
following data columns: RIAGENDR, RIDRETH1, BMXWT, RIDAGEYR}
}
\value{
Estimated daily creatinine excretion in mg creatinine/day
}
\description{
Similar to creatinine() but adds random noise to the result
}
\examples{
# daily_creatinine <- creatinine(newdata)

}
