% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readNHANES.R
\name{readNHANES}
\alias{readNHANES}
\title{readNHANES}
\usage{
readNHANES(
  codes_file,
  data_path = NULL,
  cohort = "newest",
  save_directory = "."
)
}
\arguments{
\item{codes_file}{Manually created xls or xlsx file containing 3 sheets:  1. NHANES
chemicals to include (with identifier, code, file, demographic, and units),
2. Associated weights, filenames, and column names associated with each phase used,
and 3. Parent-metabolite map containing chemical identifiers and molecular weights.}

\item{data_path}{String providing the path to the raw data.  Is "rawData",
the direct output from running get_NHANES_data would be "rawData", if left NULL, resulting
in e.g. ./rawData/1999-2000  If the directory is a subdirectory of your working one,
include "./" (e.g. "./test" will look in "./test/rawData")}

\item{cohort}{NHANES cycle/phase/cohort to analyze within the codes file. Default is to use the most recent
cycle. Other options include "oldest", "all"}

\item{save_directory}{String providing the directory in which to save the NHANES data files.  If left as the default,
NULL, it will save to ./rawData.  Otherwise, it will save to save_directory/rawData.}
}
\value{
Measured: a data frame containing a row for each metabolite with columns
providing the chemical identifiers, log geometric means, subpopulation,
and other data carried on from the codes file.
pred.data: a data frame containing a row for each parent chemical, identifiers,
MW, and a near field exposure indicator (obtained from a use database).
Uses:      a data frame with indicator columns for various chemical uses
mapping:   the mapping information in the 3rd sheet of the codes file, now matched
on the chemicals in Measured
}
\description{
Read the NHANES raw data and calls getNHANESquantile to obtain geometric mean
of concentration for each metabolite.  Matches these to those in the metabolite
map and some other objects needed for the next step.
}
\examples{
# readNHANES("NHANEScodes.xlsx", "rawData")

}
