% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examine_distribution.R
\name{doplots}
\alias{doplots}
\title{doplots}
\usage{
doplots(j, dsgn)
}
\arguments{
\item{j}{Row number of dsgn}

\item{dsgn}{Measured (output from running the readNHANES function)
with the Prabove column, sorted by this column.}
}
\value{
outplots a list with length equal to the number of chemicals,
each list element being a list with 5 components.
1. plot: ???
2. Leftseg: left interval of curve around expected
mean of distribution
3. Rightseg:right interval of curve around expected
mean of distribution
4. se.robust: ???
5. se.model: ???
Measured with 3 columns added.
1. Prabove: fraction of measurements above the LOD.
2. Rightseg: right interval of curve around expected
mean of distribution
3. Leftseg: left interval of curve around expected
mean of distribution
}
\description{
Derivative of the loglike function.  Obtain the gradient
to help in calculating variance estimates.
}
